Split out TargetControl property to ContentSizer control from SplitBase
